Supermarket

+++ app/src/main/java/com/example/finalproject/Supermarket.java
@@ -3,14 +3,14 @@ package com.example.finalproject;
 public class Supermarket implements ISupermarket {
     int superID;
     int numOfRows;
-    int lon;
-    int lat;
+    double lon;
+    double lat;
     ProductsList products;
 
     public Supermarket() {
     }
 
-    public Supermarket(int superID, int numOfRows, int lon, int lat, ProductsList products) {
+    public Supermarket(int superID, int numOfRows, double lon, double lat, ProductsList products) {
         this.superID = superID;
         this.numOfRows = numOfRows;
         this.lon = lon;
@@ -36,23 +36,26 @@ public class Supermarket implements ISupermarket {
         return this;
     }
 
-    public int getLon() {
+
+
+    public double getLon() {
         return lon;
     }
 
-    public Supermarket setLon(int lon) {
+    public Supermarket setLon(double lon) {
         this.lon = lon;
         return this;
     }
 
-    public int getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public Supermarket setLat(int lat) {
+    public Supermarket setLat(double lat) {
         this.lat = lat;
         return this;
     }
+
 
     public ProductsList getProducts() {
         return products;
